HR-Insights_ChatwithDB/requirements.txt
faker
python-dotenv
streamlit
google-generativeai